pyproject.toml
[project]
name = "ev-charging-manager"
version = "0.1.0"
description = "EV Charging Manager - Home Assistant custom integration"
readme = "README.md"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "pytest-homeassistant-custom-component>=0.13.0",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
